Fix case of duplicate column names in deconstruct
Fixes #3. The issue here was that after detecting the final column names
and types from the first element of a `Vector{T}`, we then passed those
names/types vectors to subsequent elements and they appended the same
names/types again. So we ended up with names * N and types * N, where N
is the number of elements in the original `Vector{T}`. The fix is pretty
easy: after we process the first element, save the names/types and use
those in the final DeconstructedRowIterator.