Refactor Strapping.construct to properly recurse (#10)
Fixes #8. I think the example in the issue originally worked, but may have gotten broken w/ a fix at some point. Going back over this code was a little tricky though since it's been a while. I basically started from scratch and built up the code again, which was a good exercise in simplifying a lot of the structure. At least locally, all tests pass plus the addition of the example from the docs, so I think this is a pretty good net improvement. The core issue was that in the old code, we weren't properly recursing in the construct methods for aggregates in the case of nested aggregates.

I'll also note for posterity that I looked into the possibility of supporting nested DictType, but I just don't think it makes a lot of sense fundamentally. The DictType is really about "slurping" all available fields into a Dict-like struct instead of the Struct/Mutable case where we know the fields that we're pulling out of the incoming table rows. I do think it could be possible for us to allow returning a Dict{IdFieldType, T} from the top-level construct instead of Vector{T} in the case of multiple return elements. Sometimes it can be handy to get the Dict back instead of the Vector{T} and then needing to key-value map that into a Dict. This would work by using the idproperty property of a struct as the Dict key entries.