Allow DictType as struct fields that slurp remaining fields
Extension to #8. Mainly adding this because @jademackay says they have
an actual use-case, though it should be noted that the functionality is
limited. I.e. you can have a DictType as a struct field, but it should
be the "last" field of the struct and it will act to "slurp" up any
remaining fields found in the input row after other struct fields have
been consumed.