Refactor Strapping.deconstruct for robustness and several fixes (#13)
* Refactor Strapping.deconstruct for robustness and several fixes

Fixes #12. After a deep review of `Strapping.deconstruct` after the
recent revamp of `Strapping.construct`, I discovered there were a couple
of bad assumptions and hence, glaring bugs in the code. For example, the
code failed to actually take into account any nesting more than 2
levels, and if it worked, it was purely coincidental. There were also
some robustness issues in the output of `Strapping.deconstruct` column
types; the last time I touched the code, I removed `Tables.schema`
entirely, but that isn't very satisfying since we are looking at each
row, so we _should_ be able to figure out the right type for each
column. The code here fixes that, utilizing `promote_type` when
necessary. We also fix a bad assumption that any ArrayType
objects/fields would be non-empty. Now, empty ArrayTypes are a bit of an
oddity by themselves anyway, which is also further addressed here.
Notably, if empty ArrayTypes are encountered, their type(s) will become
`Union{T, Missing}`. Conceptually this is like taking the root object of
type `T` and knowing that in the usual case, we'll produce a single row
per object. In the case of collection fields, non-collection fields will
be repeated for each collection element, but if the collection field
itself is empty, we'll still produce a single row, but the collection
field(s) will produce `missing`.

That's a decent point to segue into another bigger idea I've had of
taking `Strapping.deconstruct` to an even more functional way of
normalizing where the presence of one or more collection fields would
result in an _iterator_ of output tables, one for the base
non-collection fields, and then one for each collection field. This
follows true normalization practices from the relational data world
where the primary object key would be the sole repeated key w/ each
collection field. We could then provide a way in `Strapping.construct`
to _provide_ multiple table inputs w/ an obvious or way to specify the
primary key and have all the collection fields processed appropriately.

Not exactly sure _how_ useful all that would be given the amount of
work, since there are lots of different workflows out there and styles
of using databases. If people think there are strong enough cases for
it, we can develop the idea further.

* Require StructTypes 1.3